Implement user authentication and management features; update README and .gitignore

+++ routers/urls.py
@@ -8,7 +8,7 @@ from datetime import datetime, timezone
 from sqlalchemy.orm import Session
 from database import SessionLocal
 from models import Urls
-
+from routers.auth import get_current_user
 
 router = APIRouter(
     prefix='/urls',
@@ -24,6 +24,8 @@ def get_db():
         db.close()
 
 db_dependency = Annotated[Session,Depends(get_db)]
+user_dependency  = Annotated[dict  , Depends(get_current_user)]
+
 now = datetime.now(timezone.utc)
 
 
@@ -47,12 +49,11 @@ class FetchShortResponseSchema(BaseModel):
         orm_mode = True
 
 
-def create_short_code(long_url: str, length: int = 8) -> str:
-    # Hash the URL
-    hash_object = hashlib.sha256(long_url.encode())
-    # Encode to base64 and decode to string
+def create_short_code(long_url: str, salt: str = '', length: int = 8) -> str:
+    # Include salt in the hash to avoid collision
+    input_str = long_url + salt
+    hash_object = hashlib.sha256(input_str.encode())
     b64_encoded = base64.urlsafe_b64encode(hash_object.digest()).decode()
-    # Return a substring for shortening
     return b64_encoded[:length]
 
 def increment_access_count(short_code:str, db):
@@ -65,22 +66,26 @@ def increment_access_count(short_code:str, db):
 
 
 @router.get('/', status_code=status.HTTP_200_OK)
-async def see_all_urls(db:db_dependency):
-    return db.query(Urls).all()
+async def see_all_urls(user:user_dependency,db:db_dependency):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,detail='you are not an existing user')
+    return db.query(Urls).filter(Urls.owner_id==user.get('id')).all()
 
 
 @router.get('/{short_code}', response_model=FetchShortResponseSchema, status_code=status.HTTP_200_OK)
 async def fetch_long_url(db:db_dependency, short_code:str):
     req_url = db.query(Urls).filter(Urls.short_code==short_code).first()
-    if req_url is not None:
-        increment_access_count(short_code,db)
-        return req_url
-    raise HTTPException(status_code= status.HTTP_404_NOT_FOUND)
+    if req_url is None:
+        raise HTTPException(status_code= status.HTTP_404_NOT_FOUND)
+    increment_access_count(short_code, db)
+    return req_url
 
 @router.post('/', status_code=status.HTTP_201_CREATED, response_model=CreatedReponseSchema)
-async def shorten_url(db: db_dependency, urlreq: CreateRequestSchema):
+async def shorten_url(user:user_dependency,db: db_dependency, urlreq: CreateRequestSchema):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,detail='you are not an existing user')
     # Check if this long URL is already shortened
-    existing_url = db.query(Urls).filter(Urls.url == urlreq.long_url).first()
+    existing_url = db.query(Urls).filter(Urls.url == urlreq.long_url).filter(Urls.owner_id==user.get('id')).first()
     if existing_url:
         return existing_url
 
@@ -98,7 +103,8 @@ async def shorten_url(db: db_dependency, urlreq: CreateRequestSchema):
         short_code=short_url,
         access_count=0,
         created_at=now,
-        updated_at=now
+        updated_at=now,
+        owner_id = user.get('id')
     )
 
     db.add(new_url)
@@ -109,19 +115,24 @@ async def shorten_url(db: db_dependency, urlreq: CreateRequestSchema):
 
 
 @router.delete('/{short_code}',status_code=status.HTTP_200_OK)
-async def delete_record(db:db_dependency, short_code:str):
-    rec_to_delete = db.query(Urls).filter(Urls.short_code==short_code).first()
+async def delete_record(user:user_dependency,db:db_dependency, short_code:str):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,detail='you are not an existing user')
+    rec_to_delete = db.query(Urls).filter(Urls.short_code==short_code).filter(Urls.owner_id==user.get('id')).first()
     if rec_to_delete is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail='No such short URL found')
-    db.query(Urls).filter(Urls.short_code==short_code).delete()
+
+    db.query(Urls).filter(Urls.short_code==short_code).filter(Urls.owner_id==user.get('id')).delete()
     db.commit()
     return {'message':'deleted successfully'}
 
 
 
 @router.put('/{short_code}',status_code = status.HTTP_200_OK)
-async def update_record(db:db_dependency, short_code:str, updatereq:CreateRequestSchema):
-    rec_to_update = db.query(Urls).filter(Urls.short_code==short_code).first()
+async def update_record(user:user_dependency,db:db_dependency, short_code:str, updatereq:CreateRequestSchema):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,detail='you are not an existing user')
+    rec_to_update = db.query(Urls).filter(Urls.short_code==short_code).filter(Urls.owner_id==user.get('id')).first()
     if rec_to_update is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail='No such short URL found')
     rec_to_update.url = updatereq.long_url
